+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 *The format of this document is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).*
 
+## [v1.1.1](https://github.com/RaspberryPiFoundation/Bits/compare/v1.1.0...v1.1.1) - 2017-12-11
+
+### Fixed
+
+- Link can accept `lightOnDark` as a prop
+- Slice can accept `white` as a prop
+
 ## [v1.1.0](https://github.com/RaspberryPiFoundation/Bits/compare/v1.0.0...v1.1.0) - 2017-12-11
 
 ### Added